refactor: move the WardEditView class under the view folder from ward_view.py to ward_management_view.py
