+++ ycms/cms/views/ward/__init__.py
@@ -1,5 +1,5 @@
 """
 This package contains views related to the ward
 """
-from .ward_management_view import WardDeleteView, WardManagementView
-from .ward_view import WardEditView, WardView
+from .ward_management_view import WardDeleteView, WardManagementView, WardEditView
+from .ward_view import WardView
